refactor(logging): collapse three logger files into two
Nuxt's auto-import was scanning all three utils/logger*.ts files and
warning that 'logger' / 'Logger' were duplicated across utils/logger.ts,
utils/logger.client.ts, and utils/logger.server.ts on every dev cycle.

Restructure:
- utils/logger.ts now contains the client shim directly (was just a
  re-export of utils/logger.client.ts after the TLA removal)
- utils/logger.server.ts moves to server/utils/logger.ts so Nuxt's
  directory convention isolates it from the app-side scanner — pino
  is auto-imported only on the server context
- utils/logger.client.ts deleted (content moved into utils/logger.ts)

The 14 server-only files now import from ~/server/utils/logger
instead of ~/utils/logger.server. Shared code (entities, services,
shared utils, composables, components, pages) keeps importing
~/utils/logger and gets the console shim.

Bundle-grep regression test still passes against the production build:
pino-pretty, pino-std-serializers, and errSerializer remain absent
from .output/public/_nuxt/.
