+++ tests/utils/logger-bundle.test.ts
@@ -3,12 +3,12 @@ import { existsSync, readdirSync, readFileSync, statSync } from 'node:fs'
 import { join } from 'node:path'
 
 /**
- * Bundle-shape regression. The env-aware switch in `utils/logger.ts` relies
- * on Vite/Rollup eliminating the dead `if (import.meta.server)` branch in the
- * client build. If that ever stops working, pino's server-only deps (the
- * `pino-pretty` transport, `pino-std-serializers`) and our own `logger.server`
- * code would silently start shipping to the browser, ballooning the bundle
- * and pulling in `node:worker_threads` references that break in browsers.
+ * Bundle-shape regression. Pino lives in `server/utils/logger.ts` and is
+ * imported only by server-runtime code. Nuxt's directory convention keeps
+ * `server/**` out of the app-side bundle, so pino's server-only deps (the
+ * `pino-pretty` transport, `pino-std-serializers`) should never reach the
+ * client. If that ever stops working, the bundle balloons and we pull in
+ * `node:worker_threads` references that break in browsers.
  *
  * The test reads the production client output at `.output/public/_nuxt/` and
  * fails if any forbidden marker appears. It auto-skips if no build is on disk
@@ -16,9 +16,9 @@ import { join } from 'node:path'
  * before `npm run test:run` for this assertion to actually fire.
  *
  * Markers chosen for stability:
- *   - `pino-pretty`     — only imported by `logger.server.ts` for dev pretty-printing
+ *   - `pino-pretty`     — only imported by `server/utils/logger.ts` for dev pretty-printing
  *   - `pino-std-serializers` — only used inside pino's server runtime
- *   - `errSerializer`   — local symbol unique to `logger.server.ts`
+ *   - `errSerializer`   — local symbol unique to `server/utils/logger.ts`
  *
  * NOT a marker: bare `pino`. That string is in `@walletconnect/logger`'s
  * client bundle (Reown ships pino in browser already) and was in the bundle
