+++ utils/logger.ts
@@ -1,27 +1,93 @@
+import { summarizeViemError } from './viem-errors'
+
 /**
  * Shared structured-logger entry point.
  *
- * `utils/logger.ts` is the import path used by code that runs in **both**
- * server and client contexts (everything under `entities/`, `services/`, and
- * the cross-cutting helpers in `utils/`). It re-exports the lightweight
- * console shim from `./logger.client` unconditionally, with no top-level
- * await and no dynamic imports — Node 24 + Nitro emit a hard
- * "unsettled top-level await" warning when a chunk's TLA doesn't resolve by
- * startup, and that path was firing once per module that imported `logger`.
+ * Used by code that runs in **both** server and client contexts (everything
+ * under `entities/`, `services/`, and the cross-cutting helpers in `utils/`).
+ * Emits via `console.*` with a flat field object so a single call produces
+ * exactly one stderr line under Node — Fargate then ships one BetterStack
+ * row per logical event. The shim does not ship pino; that would balloon
+ * the client bundle for no benefit, since browser logs go to DevTools and
+ * not to BetterStack.
  *
  * Server-only modules (`server/api/**`, `server/middleware/**`,
- * `server/plugins/**`, `server/utils/**`) import `~/utils/logger.server`
- * directly to get the pino instance with JSON output, level filtering, and
- * `pino-pretty` in dev. They never need the cross-context guard.
+ * `server/plugins/**`, `server/utils/**`) import from `~/server/utils/logger`
+ * instead to get the pino instance with JSON output, level filtering, and
+ * `pino-pretty` in dev.
  *
- * Why the shim is sufficient for code that runs on the server via the
- * shared path: the client shim emits via `console.warn(prefix, msg, fields)`
- * with a flat ≤8-field object. Node's `util.inspect` keeps that on a single
- * line (well under the default `breakLength` of 128), and Fargate captures
- * one stderr line per `console.*` call — i.e. one BetterStack row per
- * logical event. The 568-row incident was specifically caused by passing
- * the raw, deeply-nested viem error object to `console.error`, which we
- * no longer do anywhere.
+ * Field shape mirrors pino: `(fields, msg)` with optional bindings via
+ * `child(bindings)`. Errors passed in `err` / `error` are summarised so we
+ * never log raw viem errors with their `abi` / `metaMessages` payloads.
  */
-export { logger } from './logger.client'
-export type { Logger } from './logger.client'
+
+type Fields = Record<string, unknown>
+type LogMethod = (fields: Fields | string, msg?: string) => void
+
+export type Logger = {
+  level: string
+  trace: LogMethod
+  debug: LogMethod
+  info: LogMethod
+  warn: LogMethod
+  error: LogMethod
+  fatal: LogMethod
+  child: (bindings: Fields) => Logger
+}
+
+const summariseErr = (value: unknown): unknown => {
+  if (value == null) return value
+  if (value instanceof Error) return summarizeViemError(value)
+  return value
+}
+
+const projectFields = (fields: Fields): Fields => {
+  const out: Fields = {}
+  for (const [key, value] of Object.entries(fields)) {
+    out[key] = key === 'err' || key === 'error' ? summariseErr(value) : value
+  }
+  return out
+}
+
+const formatPrefix = (fields: Fields): string => {
+  const ctx = typeof fields.ctx === 'string' ? fields.ctx : undefined
+  const chain = typeof fields.chain === 'string' ? fields.chain : undefined
+  const chainId = typeof fields.chainId === 'number' ? fields.chainId : undefined
+  const chainTag = chain != null ? ` (${chain}${chainId != null ? `:${chainId}` : ''})` : ''
+  return ctx ? `[${ctx}]${chainTag}` : `[log]${chainTag}`
+}
+
+type ConsoleMethod = 'debug' | 'info' | 'warn' | 'error'
+
+const make = (bindings: Fields): Logger => {
+  // Look up console.* at call time, not module load, so test spies and any
+  // userland console replacement (Sentry breadcrumbs etc.) take effect. The
+  // eslint-disable directives on debug/info are intentional: this shim is the
+  // single sanctioned bridge from our structured-logging API to browser
+  // DevTools, and we want trace/debug/info entries to remain visible at their
+  // natural verbosity rather than collapsing onto warn.
+  const emit = (method: ConsoleMethod) =>
+    (fieldsOrMsg: Fields | string, msg?: string): void => {
+      if (typeof fieldsOrMsg === 'string') {
+        // eslint-disable-next-line no-console
+        console[method](formatPrefix(bindings), fieldsOrMsg)
+        return
+      }
+      const fields = projectFields({ ...bindings, ...fieldsOrMsg })
+      // eslint-disable-next-line no-console
+      console[method](formatPrefix(fields), msg ?? '', fields)
+    }
+
+  return {
+    level: 'debug',
+    trace: emit('debug'),
+    debug: emit('debug'),
+    info: emit('info'),
+    warn: emit('warn'),
+    error: emit('error'),
+    fatal: emit('error'),
+    child: (extra: Fields) => make({ ...bindings, ...extra }),
+  }
+}
+
+export const logger: Logger = make({})
